# api.swagger.yaml
swagger: "2.0"
info:
  title: Blog Server.
  description: Entry point for the application.
  version: 1.0.0
  contact:
    name: HardDie
    url: https://github.com/HardDie
    email: oleg1995sysoev@yandex.ru
...
basePath: /
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/v1/auth/login:
    post:
      summary: Login form
      operationId: Auth_Login
      responses:
        "200":
          description: A successful response.
          schema:
            type: object
            properties: {}
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/gatewayLoginRequest'
      tags:
        - Auth
  /api/v1/auth/logout:
    post:
      summary: Logout
      operationId: Auth_Logout
      responses:
        "200":
          description: A successful response.
          schema:
            type: object
            properties: {}
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties: {}
      tags:
        - Auth
  /api/v1/auth/register:
    post:
      summary: Registration form
      operationId: Auth_Register
      responses:
        "200":
          description: A successful response.
          schema:
            type: object
            properties: {}
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/gatewayRegisterRequest'
      tags:
        - Auth
  /api/v1/auth/user:
    get:
      summary: Getting information about the current user
      operationId: Auth_User
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayUserResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      tags:
        - Auth
...
definitions:
...
  gatewayLoginRequest:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
...
  gatewayRegisterRequest:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
      displayedName:
        type: string
      invite:
        type: string
  gatewayUserResponse:
    type: object
    properties:
      id:
        type: string
        format: int64
      username:
        type: string
      displayedName:
        type: string
      email:
        type: string
      invitedByUserId:
        type: string
        format: int64
      createdAt:
        type: string
        format: date-time
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
securityDefinitions:
  bearer:
    type: apiKey
    description: 'Authentication token, prefixed by Bearer: Bearer <token>'
    name: Authorization
    in: header
security:
  - bearer: []